OD-18887 fix: merge of related checkouts

+++ server/src/main/groovy/ish/oncourse/server/deduplication/ContactMergeService.groovy
@@ -308,12 +308,12 @@ class ContactMergeService {
 
         context.deleteObjects(b.customFields)
 
-        if(b.abandonedCarts) {
+        if(b.abandonedCarts && !b.abandonedCarts.empty) {
             if (a.abandonedCarts.isEmpty() || a.abandonedCarts.first().createdOn.before(b.abandonedCarts.first().createdOn)) {
-                b.abandonedCarts.first().setPayer(a)
-
                 if(!a.abandonedCarts.empty)
                     context.deleteObjects(a.abandonedCarts)
+
+                b.abandonedCarts.first().setPayer(a)
             }
         }
